Migrate `evalStory` handler to Stories Saga
